requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
cache==1.0.3
cachelib==0.9.0
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
eventlet==0.36.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.7
lml==0.1.0
MarkupSafe==2.1.5
passlib==1.7.4
pdfkit==1.0.0
prompt-toolkit==3.0.43
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
setuptools==75.5.0
six==1.16.0
SQLAlchemy==2.0.25
texttable==1.7.0
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
